Project name changed: MCSkyBot -> JDSkyBot
This project will now be focused on creating utilities for Discord bots made with Java in general, not only oriented to Minecraft, as that was very specific. This name (JDS) also resembles the name of the API it uses (JDA). The full name of the project would be something like Java-Discord SkyBot