Make JDSkyBot modular. Features are now optional
I want to make JDSky as modular as possible. Jar size is not really an issue on this project, given that JDA is currently being shaded onto it, and that makes a jar of around 15MB. However, as it happens with JSky, JDSky probably won't shade any dependency, giving the developer the option to use whatever version of JDA they want. Also, this would remove the need to update JDSky every time JDA releases an update, or I update JSky... As long as it still works, of course